:recycle: - Combine zaak related services.
