+++ frontend/zac-ui/libs/features/kownsl/src/lib/approval/approval.service.ts
@@ -3,7 +3,6 @@ import { ApplicationHttpClient } from '@gu/services';
 import { Observable } from 'rxjs';
 import { ApprovalForm } from '../../models/approval-form';
 import { HttpResponse } from '@angular/common/http';
-import { Zaak } from '@gu/models';
 
 @Injectable({
   providedIn: 'root'
@@ -18,10 +17,6 @@ export class ApprovalService {
       observe: 'response' as 'response'
     }
     return this.http.Get<any>(endpoint, options);
-  }
-
-  getZaakDetail(bronorganisatie: string, identificatie: string): Observable<Zaak> {
-    return this.http.Get<Zaak>(encodeURI(`/api/core/cases/${bronorganisatie}/${identificatie}`));
   }
 
   postApproval(formData: ApprovalForm, uuid:string): Observable<any> {
