--- frontend/zac-ui/libs/features/zaak-detail/src/lib/documenten/documenten.interface.ts
@@ -1,34 +1,0 @@
-export interface Document {
-  auteur: string;
-  beschrijving: string;
-  bestandsnaam: string;
-  bestandsomvang: number;
-  identificatie: string;
-  informatieobjecttype: InformatieObjectType;
-  locked: boolean;
-  readUrl: string;
-  title: string;
-  url: string;
-  vertrouwelijkheidaanduiding: string;
-  versie: number,
-  writeUrl: string;
-}
-
-interface InformatieObjectType {
-  omschrijving: string;
-  url: string;
-}
-
-export interface DocumentUrls {
-  writeUrl?: string;
-  deleteUrl?: string;
-  drcUrl?: string;
-  id?: string;
-}
-
-export interface ReadWriteDocument {
-  magicUrl: string;
-  deleteUrl: string;
-  drcUrl: string;
-  purpose: string;
-}
