+++ frontend/zac-ui/libs/shared/data-access/models/accounts/user-permissions.ts
@@ -1,0 +1,6 @@
+import {Permission} from "../zaken/permission";
+
+export interface UserPermission {
+  permissions: Permission[],
+  username: string,
+}
