+++ frontend/zac-ui/libs/shared/data-access/models/index.ts
@@ -6,9 +6,11 @@ export * from './ui/menu';
 
 // Accounts
 export * from './accounts/user';
+export * from './accounts/user-permissions';
 
 // Zaken
 export * from './zaken/assignee'
 export * from './zaken/document'
+export * from './zaken/permission'
 export * from './zaken/task'
 export * from './zaken/zaak'
