+++ frontend/zac-ui/libs/shared/data-access/models/zaken/document.ts
@@ -1,16 +1,3 @@
-export interface ShortDocument {
-  bestandsgrootte: string;
-  documentType: string;
-  readUrl: string;
-  titel: string;
-  vertrouwelijkheidaanduiding: string;
-}
-
-interface InformatieObjectType {
-  omschrijving: string;
-  url: string;
-}
-
 export interface Document {
   auteur: string;
   beschrijving: string;
@@ -23,5 +10,33 @@ export interface Document {
   title: string;
   url: string;
   vertrouwelijkheidaanduiding: string;
+  versie: number,
   writeUrl: string;
+}
+
+export interface DocumentUrls {
+  writeUrl?: string;
+  deleteUrl?: string;
+  drcUrl?: string;
+  id?: string;
+}
+
+export interface ReadWriteDocument {
+  magicUrl: string;
+  deleteUrl: string;
+  drcUrl: string;
+  purpose: string;
+}
+
+export interface ShortDocument {
+  bestandsgrootte: string;
+  documentType: string;
+  readUrl: string;
+  titel: string;
+  vertrouwelijkheidaanduiding: string;
+}
+
+interface InformatieObjectType {
+  omschrijving: string;
+  url: string;
 }
