+++ frontend/zac-ui/libs/shared/data-access/services/src/index.ts
@@ -1,2 +1,3 @@
-export * from './lib/shared-data-access-services.module';
 export * from './lib/http/http-client';
+export * from './lib/shared-data-access-services.module';
+export * from './lib/zaak/zaak.services'
